remove some excess arguments?
I don't know if this is a good fix, but it fixed my problem. There were exceptions
stating that the .read() and .write() functions each wanted 4 arguments, not 5. removing
the last one in each case got me past that, and the xsusbprg and xstest scripts were able
to function. Whether this is the right sort of fix or why I'm seeing this problem, I didn't
really investigate.